Fix web compilation by isolating native database executor
