+++ bulletin_app/lib/data/local/app_database_executor_stub.dart
@@ -1,0 +1,4 @@
+import 'package:drift/drift.dart';
+
+QueryExecutor createExecutor() =>
+    throw UnsupportedError('No database executor available for this platform.');
