+++ bulletin_app/lib/data/local/app_database_executor_web.dart
@@ -1,0 +1,4 @@
+import 'package:drift/drift.dart';
+import 'package:drift/web.dart';
+
+QueryExecutor createExecutor() => WebDatabase('bulletins_db');
